ISTA557/back-propagation-samfreitas1996/requirements.txt
numpy >= 1.17
scipy >= 1.3
pytest-timeout >= 1.3
